+++ rollup.config.js
@@ -1,8 +1,10 @@
-import babel from 'rollup-plugin-babel';
+import babel from '@rollup/plugin-babel';
 
 // eslint-disable-next-line import/no-anonymous-default-export
 export default {
   input: 'src/index.js',
-  plugins: [babel()],
+  plugins: [babel({
+    babelHelpers: 'bundled'
+  })],
   output: {sourcemap: true}
 };