+++ rollup.config.tests.js
@@ -1,11 +1,13 @@
-import babel from 'rollup-plugin-babel';
+import babel from '@rollup/plugin-babel';
 import multiEntry from 'rollup-plugin-multi-entry';
 import async from 'rollup-plugin-async';
 
 // eslint-disable-next-line import/no-anonymous-default-export
 export default {
   input: 'test/**/*.test.js',
-  plugins: [async(), babel(), multiEntry()],
+  plugins: [async(), babel({
+    babelHelpers: 'bundled'
+  }), multiEntry()],
   output: {
     format: 'umd',
     file: 'build/tests-bundle.js',